Rm circular deps & rm x dep between main/renderer
Just simplifying out the import graph reducing and removing the cross dependance loads un-needed code to be loaded into memory. But this is probably treeshaken anyway
