+++ src/common/AmongUsState.ts
@@ -1,0 +1,29 @@
+
+export interface AmongUsState {
+	gameState: GameState;
+	oldGameState: GameState;
+	lobbyCode: string;
+	players: Player[];
+}
+export interface Player {
+	ptr: number;
+	id: number;
+	name: string;
+	colorId: number;
+	hatId: number;
+	petId: number;
+	skinId: number;
+	disconnected: boolean;
+	isImpostor: boolean;
+	isDead: boolean;
+	taskPtr: number;
+	objectPtr: number;
+	isLocal: boolean;
+
+	x: number;
+	y: number;
+	inVent: boolean;
+}
+export enum GameState {
+	LOBBY, TASKS, DISCUSSION, MENU, UNKNOWN
+}
