+++ src/common/ISettings.d.ts
@@ -1,0 +1,16 @@
+
+export interface ISettings {
+	alwaysOnTop: boolean;
+	microphone: string;
+	speaker: string;
+	pushToTalk: boolean;
+	serverURL: string;
+	pushToTalkShortcut: string;
+	deafenShortcut: string;
+	offsets: {
+		version: string;
+		data: string;
+	},
+	hideCode: boolean;
+	stereoInLobby: boolean;
+}
