+++ src/renderer/MicrophoneSoundBar.tsx
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useState } from 'react'
-import { SettingsContext } from './App';
+import { SettingsContext } from './contexts';
 
 const TestMicrophoneButton = function() {
     const [{ microphone }] = useContext(SettingsContext)
