+++ src/renderer/contexts.tsx
@@ -1,0 +1,10 @@
+import React, { createContext } from 'react';
+import { AmongUsState } from "../common/AmongUsState";
+import { ISettings } from "../common/ISettings";
+
+
+export const GameStateContext = createContext<AmongUsState>({} as AmongUsState);
+export const SettingsContext = createContext<[ISettings, React.Dispatch<{
+	type: "set" | "setOne";
+	action: ISettings | [string, any];
+}>]>(null as any);
